Swag

+++ index.php
@@ -9,7 +9,7 @@
         $url="http://api.openweathermap.org/data/2.5/weather?q=Charlotte,US&units=metric&cnt=7&lang=en";
         $json=file_get_contents($url);
         $data=json_decode($json,true);
-        $temp = floatval($data['main']['temp']);
+        $temp = intval($data['main']['temp']);
         $temp = $temp* 2 + 32.00;
         $end = $data['weather'][0]['description']." and the temperature is ".$temp;
     }
